+++ src/dtkMath/dtkMath.h
@@ -4,9 +4,9 @@
  * Copyright (C) 2008 - Thibaud Kloczko, Inria.
  * Created: Tue Jun  8 15:37:50 2010 (+0200)
  * Version: $Id$
- * Last-Updated: Wed Mar  9 15:18:54 2011 (+0100)
+ * Last-Updated: Tue Mar 15 14:19:00 2011 (+0100)
  *           By: Thibaud Kloczko
- *     Update #: 170
+ *     Update #: 176
  */
 
 /* Commentary: 
@@ -22,6 +22,8 @@
 
 #include "dtkMathExport.h"
 
+#include <inttypes.h>
+
 DTKMATH_EXPORT void dtkBubbleSort(unsigned int indices[], int size);
 
 DTKMATH_EXPORT bool dtkAlmostEqualRelative( float A,  float B,  float maxRelativeError);
@@ -33,11 +35,11 @@ DTKMATH_EXPORT bool dtkAlmostEqualRelative2(double A, double B, double maxRelati
 DTKMATH_EXPORT bool dtkAlmostEqualRelativeOrAbsolute( float A,  float B,  float maxRelativeError,  float maxAbsoluteError);
 DTKMATH_EXPORT bool dtkAlmostEqualRelativeOrAbsolute(double A, double B, double maxRelativeError, double maxAbsoluteError);
 
-DTKMATH_EXPORT bool dtkAlmostEqualUlpsSimple( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkAlmostEqualUlpsSimple(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkAlmostEqualUlpsSimple( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkAlmostEqualUlpsSimple(double A, double B, int_least64_t maxUlps);
 
-DTKMATH_EXPORT bool dtkAlmostEqual2sComplement( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkAlmostEqual2sComplement(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkAlmostEqual2sComplement( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkAlmostEqual2sComplement(double A, double B, int_least64_t maxUlps);
 
 DTKMATH_EXPORT bool dtkIsInfinite( float A);
 DTKMATH_EXPORT bool dtkIsInfinite(double A);
@@ -45,26 +47,26 @@ DTKMATH_EXPORT bool dtkIsInfinite(double A);
 DTKMATH_EXPORT bool dtkIsNan( float A);
 DTKMATH_EXPORT bool dtkIsNan(double A);
 
-DTKMATH_EXPORT           int dtkSign( float A);
-DTKMATH_EXPORT long long int dtkSign(double A);
+DTKMATH_EXPORT int_least32_t dtkSign( float A);
+DTKMATH_EXPORT int_least64_t dtkSign(double A);
 
-DTKMATH_EXPORT bool dtkAlmostEqualUlps( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkAlmostEqualUlps(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkAlmostEqualUlps( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkAlmostEqualUlps(double A, double B, int_least64_t maxUlps);
 
-DTKMATH_EXPORT void dtkPrintNumber( float f,           int offset);
-DTKMATH_EXPORT void dtkPrintNumber(double f, long long int offset);
+DTKMATH_EXPORT void dtkPrintNumber( float f, int_least32_t offset);
+DTKMATH_EXPORT void dtkPrintNumber(double f, int_least64_t offset);
 
-DTKMATH_EXPORT bool dtkLesserThanUlps( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkLesserThanUlps(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkLesserThanUlps( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkLesserThanUlps(double A, double B, int_least64_t maxUlps);
 
-DTKMATH_EXPORT bool dtkLesserThanOrAlmostEqualUlps( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkLesserThanOrAlmostEqualUlps(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkLesserThanOrAlmostEqualUlps( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkLesserThanOrAlmostEqualUlps(double A, double B, int_least64_t maxUlps);
 
-DTKMATH_EXPORT bool dtkGreaterThanUlps( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkGreaterThanUlps(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkGreaterThanUlps( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkGreaterThanUlps(double A, double B, int_least64_t maxUlps);
 
-DTKMATH_EXPORT bool dtkGreaterThanOrAlmostEqualUlps( float A,  float B,           int maxUlps);
-DTKMATH_EXPORT bool dtkGreaterThanOrAlmostEqualUlps(double A, double B, long long int maxUlps);
+DTKMATH_EXPORT bool dtkGreaterThanOrAlmostEqualUlps( float A,  float B, int_least32_t maxUlps);
+DTKMATH_EXPORT bool dtkGreaterThanOrAlmostEqualUlps(double A, double B, int_least64_t maxUlps);
 
 #include "dtkMath.tpp"
